Mode codemirror.js under app/javascript/packs
Meld codemirror_fluentd.js into codemirror.js.
